Centralize CDX field formatting code in CdxFormat
The fact we have to add overloads to CdxFields.format() highlights that the way the field formatting code is split between the CdxFields and CdxFormat classes is awkward and confusing. Let's just move it all to CdxFormat.

#74
