+++ test/org/netpreserve/jwarc/cdx/CdxFormatTest.java
@@ -27,9 +27,10 @@ public class CdxFormatTest {
 
     @Test
     public void testDigestUnchanged() throws Exception {
-        CdxFormat cdxFormat = CdxFormat.CDX11;
-        cdxFormat.setDigestUnchanged(true); //We want the digest as is.
-        String payloadDigest="sha256:b04af472c47a8b1b5059b3404caac0e1bfb5a3c07b329be66f65cfab5ee8d3f3";    
+        CdxFormat cdxFormat = new CdxFormat.Builder()
+                .digestUnchanged() // We want the digest as is.
+                .build();
+        String payloadDigest="sha256:b04af472c47a8b1b5059b3404caac0e1bfb5a3c07b329be66f65cfab5ee8d3f3";
                 
         HttpResponse httpResponse = new HttpResponse.Builder(404, "Not Found")
                 .body(MediaType.HTML, new byte[0])
